requirements.txt
jupyter  
matplotlib
graphviz
pillow
ipython